+++ pkg/stdlib/arrays/shift.go
@@ -25,7 +25,7 @@ func Shift(_ context.Context, args ...core.Value) (core.Value, error) {
 
 	arr := args[0].(*values.Array)
 
-	length := int(arr.Length())
+	length := int64(arr.Length())
 	result := values.NewArray(length)
 
 	arr.ForEach(func(value core.Value, idx int) bool {
